ble: Add explicit uint32_t cast to HUBBLE_MIN() call
HUBBLE_MIN(time_remaining, UINT32_MAX) evaluates to uint64_t, so
returning it from hubble_ble_advertise_expiration_get() implicitly
converts it down to uint32_t.

The conversion is now explicit. This is safe as HUBBLE_MIN() clamps
time_remaining down to UINT32_MAX before returning.

+++ src/hubble_ble.c
@@ -116,5 +116,5 @@ uint32_t hubble_ble_advertise_expiration_get(void)
 	uint64_t time_in_current_period = time_ms % rotation_period_ms;
 	uint64_t time_remaining = rotation_period_ms - time_in_current_period;
 
-	return HUBBLE_MIN(time_remaining, UINT32_MAX);
+	return (uint32_t)HUBBLE_MIN(time_remaining, UINT32_MAX);
 }
